refactor: extract Partners page into section components
- Split 1530-line Partners.tsx into 6 focused section components
- Extract partner data to data/partners-data.ts (312 lines)
- Remove barrel files (atoms/index.ts, molecules/index.ts, organisms/index.ts)
- Update all imports to use direct paths instead of barrel re-exports
- New components: hero-section, philosophy-section, ecosystem-section,
  partner-card, partner-form, cta-section
