+++ App.tsx
@@ -11,8 +11,8 @@ import Home from './pages/Home';
 import { JobDetail } from './pages/JobDetail';
 import { Jobs } from './pages/Jobs';
 import Method from './pages/Method';
-import { Partners } from './pages/Partners';
 import { Podcast } from './pages/Podcast';
+import { Partners } from './pages/partners';
 import { Resources } from './pages/Resources';
 import Services from './pages/Services';
 import { Stories } from './pages/Stories';
