--- components/atoms/index.ts
@@ -1,4 +1,0 @@
-export { Badge } from './Badge';
-export { Button } from './Button';
-export { Input } from './Input';
-export { default as NumberCircle } from './NumberCircle';
