--- components/molecules/index.ts
@@ -1,8 +1,0 @@
-export { AccentCard } from './AccentCard';
-export { Card } from './Card';
-export { PillarCard } from './PillarCard';
-export { ScopeCard } from './ScopeCard';
-export { ServiceCard } from './ServiceCard';
-export { TeamMemberCard } from './TeamMemberCard';
-export { TestimonialCard } from './TestimonialCard';
-export { ValueCard } from './ValueCard';
