--- components/organisms/index.ts
@@ -1,3 +1,0 @@
-export { Container } from './Container';
-export { Section } from './Section';
-export { SectionHeader } from './SectionHeader';
