+++ components/sections/home/ComparatifSection.tsx
@@ -2,8 +2,9 @@ import { Check, X } from 'lucide-react';
 import type React from 'react';
 import { css } from 'styled-system/css';
 import { center, flex, grid } from 'styled-system/patterns';
-import { Button } from '../../atoms';
-import { Container, Section } from '../../organisms';
+import { Button } from '../../atoms/Button';
+import { Container } from '../../organisms/Container';
+import { Section } from '../../organisms/Section';
 
 const OCOBO_BENEFITS = [
 	"Commence par la stratégie, pas par l'outil",
