+++ components/sections/method/ScopeSection.tsx
@@ -11,8 +11,9 @@ import { css } from 'styled-system/css';
 import { grid } from 'styled-system/patterns';
 import { ThemeColor } from '../../../types';
 import { UnifiedBowtie } from '../../illustrations';
-import { ScopeCard } from '../../molecules';
-import { Container, Section } from '../../organisms';
+import { ScopeCard } from '../../molecules/ScopeCard';
+import { Container } from '../../organisms/Container';
+import { Section } from '../../organisms/Section';
 
 const scopeData = [
 	{
