+++ components/sections/services/HeroSection.tsx
@@ -2,7 +2,8 @@ import { ChevronDown } from 'lucide-react';
 import type React from 'react';
 import { css } from 'styled-system/css';
 import { center, flex } from 'styled-system/patterns';
-import { Badge, Button } from '../../atoms';
+import { Badge } from '../../atoms/Badge';
+import { Button } from '../../atoms/Button';
 import { ArchitecturalGrid } from '../../illustrations';
 import HeroSplit from '../../layout/HeroSplit';
