+++ components/sections/services/MethodeSection.tsx
@@ -2,7 +2,7 @@ import { BarChart3, Cog, GraduationCap, Layout } from 'lucide-react';
 import type React from 'react';
 import { css } from 'styled-system/css';
 import { center, grid } from 'styled-system/patterns';
-import { SectionHeader } from '../../organisms';
+import { SectionHeader } from '../../organisms/SectionHeader';
 
 const STEPS = [
 	{
