+++ pages/ArticleDetail.tsx
@@ -14,8 +14,8 @@ import type React from 'react';
 import { useEffect, useState } from 'react';
 import { Link, useParams } from 'react-router';
 import { css } from 'styled-system/css';
-import { center, flex, grid, hstack, vstack } from 'styled-system/patterns';
-import { Badge } from '../components/atoms';
+import { flex, grid, hstack } from 'styled-system/patterns';
+import { Badge } from '../components/atoms/Badge';
 
 export const ArticleDetail: React.FC = () => {
 	const { slug: _slug } = useParams();
